requirements.txt
alabaster==0.7.16
astroid==3.1.0
attrs==23.2.0
Babel==2.14.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
dill==0.3.8
docutils==0.20.1
fiona==1.9.6
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
geopandas==0.14.3
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
imagesize==1.4.1
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mccabe==0.7.0
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pandas==2.2.1
parsimonious==0.10.0
platformdirs==4.2.0
psycopg2-binary==2.9.9
Pygments==2.17.2
pylint==3.1.0
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.4.0
shapely==2.0.3
six==1.16.0
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinx-adc-theme==0.1.7
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
SQLAlchemy==2.0.28
tomlkit==0.12.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
